# artifacts/api-server/src/routes/tutor.ts
import { Router, type IRouter } from "express";
import { eq } from "drizzle-orm";
import { db, lecturesTable } from "@workspace/db";
import { AskTutorBody, AskTutorResponse } from "@workspace/api-zod";
import { chatText, chatJson, FAST_MODEL } from "../lib/ai";

const router: IRouter = Router();

router.get("/tutor/suggestions/:lectureId", async (req, res): Promise<void> => {
  const lectureId = Number(req.params.lectureId);
  if (!Number.isFinite(lectureId)) {
    res.status(400).json({ error: "invalid lectureId" });
    return;
  }
  const [lecture] = await db
    .select()
    .from(lecturesTable)
    .where(eq(lecturesTable.id, lectureId));
  if (!lecture) {
    res.status(404).json({ error: "lecture not found" });
    return;
  }

  try {
    const out = await chatJson<{ questions: string[] }>(
      'You are an encouraging college conceptual-physics tutor. Reply as strict JSON of the form {"questions": string[]} with NO other keys.',
      `From the lecture below, generate 6 short, concrete starter questions a student might want to ask after reading it. Cover every major idea in the reading (not just the first one). Each question must be one sentence, under ~18 words, in the student's voice (e.g. "Why does ...?", "Can you show me ...?", "What's the difference between ...?").\n\nLECTURE TITLE: ${lecture.title}\n\nLECTURE BODY:\n"""\n${lecture.body}\n"""`,
      FAST_MODEL,
    );
    const questions = Array.isArray(out?.questions)
      ? out.questions.filter((q) => typeof q === "string" && q.trim().length > 0).slice(0, 8)
      : [];
    res.json({ questions });
  } catch {
    res.json({ questions: [] });
  }
});

router.post("/tutor/ask", async (req, res): Promise<void> => {
  const parsed = AskTutorBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const { message, selectedLectureText, context } = parsed.data;

  const sys =
    "You are an encouraging college conceptual-physics tutor. Explain step by step, use clear physical intuition and concrete everyday examples, and name the relevant concepts (forces, energy, momentum, fields, waves, entropy, relativity, quantum behavior) where helpful. Favor conceptual understanding over heavy algebra, but use simple equations like F=ma or E=mc² when they illuminate the idea. Keep replies short (3-6 sentences) unless the student asks for more detail. Never just give the answer — guide them.";
  const blocks: string[] = [];
  if (selectedLectureText) {
    blocks.push(
      `Context from the lecture the student is reading:\n"""\n${selectedLectureText}\n"""`,
    );
  }
  if (context) {
    blocks.push(
      `The student is in a practice session. Here is the problem they just worked, their answer, and the feedback they're discussing:\n"""\n${context}\n"""`,
    );
  }
  blocks.push(`Student question: ${message}`);
  const user = blocks.join("\n\n");

  let text = "";
  try {
    text = await chatText(sys, user);
  } catch {
    text =
      "I'm having trouble reaching the tutor service right now. Try again in a moment, and consider re-reading the relevant section of the lecture.";
  }
  res.json(AskTutorResponse.parse({ text, audioUrl: null }));
});

export default router;