# lib/api-zod/src/generated/types/practiceSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PracticeSession {
  id: number;
  tutorEnabled: boolean;
  difficulty: number;
  /** @nullable */
  weekNumber: number | null;
  /** @nullable */
  topicId?: number | null;
  /** @nullable */
  assignmentId?: number | null;
  focusOnWeaknesses?: boolean;
}