# lib/api-zod/src/generated/types/tutorAskInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TutorAskInput {
  /** @nullable */
  sessionId?: number | null;
  message: string;
  /** @nullable */
  selectedLectureText?: string | null;
  /**
     * Extra grounding context, e.g. the current practice problem, the student's answer, and the feedback they're asking about.
     * @nullable
     */
  context?: string | null;
  /** @nullable */
  lectureId?: number | null;
  /** @nullable */
  problemId?: number | null;
}